services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
redis==5.2.1
psutil==6.1.1
python-dotenv==1.0.1
httpx==0.28.1
